+++ Editor/Scripts/GltfImporter.cs
@@ -17,6 +17,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using UnityEditor;
 #if UNITY_2020_2_OR_NEWER
 using UnityEditor.AssetImporters;
@@ -24,6 +25,7 @@ using UnityEditor.AssetImporters;
 using UnityEditor.Experimental.AssetImporters;
 #endif
 using UnityEngine;
+using UnityEngine.Rendering;
 using Object = UnityEngine.Object;
 
 namespace GLTFast.Editor {
@@ -31,6 +33,9 @@ namespace GLTFast.Editor {
     [ScriptedImporter(1,new [] {"gltf","glb"})] 
     public class GltfImporter : ScriptedImporter {
 
+        [SerializeField]
+        EditorImportSettings editorImportSettings;
+        
         [SerializeField]
         ImportSettings importSettings;
         
@@ -68,8 +73,13 @@ namespace GLTFast.Editor {
                 logger
                 );
 
-            if (importSettings == null) {
+            if (editorImportSettings == null) {
                 // Design-time import specific settings
+                editorImportSettings = new EditorImportSettings();
+            }
+            
+            if (importSettings == null) {
+                // Design-time import specific changes to default settings
                 importSettings = new ImportSettings {
                     // Avoid naming conflicts by default
                     nodeNameMethod = ImportSettings.NameImportMethod.OriginalUnique
@@ -123,6 +133,9 @@ namespace GLTFast.Editor {
                 var meshes = m_Gltf.GetMeshes();
                 if (meshes != null) {
                     foreach (var mesh in meshes) {
+                        if (editorImportSettings.generateSecondaryUVSet && !HasSecondaryUVs(mesh)) {
+                            Unwrapping.GenerateSecondaryUVSet(mesh);
+                        }
                         AddObjectToAsset(ctx, $"meshes/{mesh.name}", mesh);
                     }
                 }
@@ -208,6 +221,11 @@ namespace GLTFast.Editor {
                 return extName;
             }
             return originalName;
+        }
+
+        static bool HasSecondaryUVs(Mesh mesh) {
+            var attributes = mesh.GetVertexAttributes();
+            return attributes.Any(attribute => attribute.attribute == VertexAttribute.TexCoord1);
         }
     }
 }