+++ ChangeLog.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - *Generate Lightmap UVs* option in the glTF import inspector lets you create a secondary texture coordinate set (similar to the Model Import Settings from other formats; #238)
 - Generic `ICodeLogger` methods that don't require a `LogCode`
+### Changed
+- Less GC due to `CollectingLogger` creating the item list on demand 
 
 ## [4.3.4] - 2021-10-26
 ### Added